requirements.txt
Flask==2.3.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.6
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
python-dotenv==1.0.1
email-validator==2.1.0.post1
